% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignmentb2torinhalvorsonRscript.R
\name{stdev}
\alias{stdev}
\title{Title stdev()}
\usage{
stdev(w)
}
\arguments{
\item{w}{a numeric vector, named 'w' in keeping with conventions for vectors}
}
\value{
The function returns the standard deviation of the vector. The mean and length of the vector are also printed as the function runs, to double-check that the values used for calculating the standard deviation are correct.
}
\description{
This function calculates the standard deviation of a vector of numbers, automatically removing any missing values. The input must be numeric, or the function will throw an error.
}
\examples{
#Calculate standard deviation of a vector
vec1 <- c(1:20)
stdev(vec1)

#Calculate standard deviation of a vector with missing values
vec2 <- c(2,3,4,5,NA,NA,12,17)
stdev(vec2)

#Calculate standard deviation of a numeric columns within a data frame
var1 <- c(1:10)
var2 <- c(3:12)
df <- data.frame(var1, var2)
stdev(df$var2)


}
